Fixed #161 Prefab reload was broken. Also removed deprecated serialization code.
