+++ Code/Engine/GameEngine/Animation/Implementation/PropertyAnimResource.cpp
@@ -1,7 +1,6 @@
 #include <GameEnginePCH.h>
 
 #include <Core/Assets/AssetFileHeader.h>
-#include <Core/WorldSerializer/ResourceHandleReader.h>
 #include <GameEngine/Animation/PropertyAnimResource.h>
 #include <GameEngine/Curves/ColorGradientResource.h>
 #include <GameEngine/Curves/Curve1DResource.h>
